+++ Changelog.md
@@ -2,6 +2,11 @@
 
 ## vNext
 
+## 0.5.6
+
+- Improved detection of incorrect usage of transient disposable object: InvalidOperationException has more information about the service being resolved (like a resolution context stack, if available).
+- Improved detection of incorrect usage of transient disposable object: added an exclusion pattern list for services that should not be checked (like some AspNetCore internal services).
+
 ## 0.5.5
 
 - `ServiceProviderFactory`: added a constructor that accepts `ExtendedServiceProviderOptions` to be used in Host Builder initialization like: